requirements.txt
Django==5.0.1
mysqlclient==2.2.1
django-crispy-forms==2.1
Pillow==10.2.0
djangorestframework==3.14.0
djoser==2.2.2
sqlalchemy==2.0.25
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.0
